Implemented a simple shell system and stack protector.
