+++ src/kernel/include/ui.h
@@ -1,0 +1,14 @@
+#ifndef _ui_h
+#define _ui_h
+
+#include <system.h>
+#include <stdlib.h>
+#include <stdio.h>
+#include <string.h>
+#include <phapi.h>
+
+void prompt();
+void welcome();
+void login();
+
+#endif
