+++ src/kernel/system/stack_protector.c
@@ -1,0 +1,23 @@
+#include <system.h>
+#include <stdlib.h>
+#include <stdio.h>
+#include <string.h>
+#include <phapi.h>
+ 
+#if UINT32_MAX == UINTPTR_MAX
+#define STACK_CHK_GUARD 0xe2dee396
+#else
+#define STACK_CHK_GUARD 0x595e9fbd94fda766
+#endif
+ 
+uintptr_t __stack_chk_guard = STACK_CHK_GUARD;
+ 
+__attribute__((noreturn))
+void __stack_chk_fail(void)
+{
+#if __STDC_HOSTED__
+	abort();
+#elif __is_photon_kernel
+	panic("Stack smashing detected", __LINE__, __FILE__);
+#endif
+}
