+++ src/kernel/system/system.c
@@ -13,110 +13,6 @@
 #include <string.h>
 #include <phapi.h>
 
-char user[20];
-char machine[30];
-
-extern mem_heap_t *kernel_heap;
-extern int mseconds;
-extern int seconds;
-extern int minutes;
-extern int hours;
-extern void *stdin;
-
-static void testing_shell(char *str) 
-{
-
-	while (true) {
-		printk("\nEnter test name> ");
-		gets(str);
-		
-		if (!strcmp(str, "exit")) {
-			break;
-		} else if (!strcmp(str, "print-stdin")) {
-			printk("%s\n", (uint8_t*)stdin);
-		} else if (!strcmp(str, "size")) {
-			printk("Sizes:\n");
-			printk("Char: %d\nInt: %d\nFloat: %d\nSize_t: %d\n", sizeof(char), sizeof(int), sizeof(float), sizeof(size_t));
-			printk("Kernel:\n");
-			printk("Start: %x\nEnd: %x\n", (size_t)&kernel_start, (size_t)&kernel_end);
-			//printk("RAM: %d MB\n", mem_size_mb);
-		} else if (!strcmp(str, "page-fault")) {
-			printk("Test fault...\n");
-			uint32_t *ptr = (uint32_t*)0xA123142300;
-			*ptr = 432;
-			printk("%d\n", *ptr);
-		} else if (!strcmp(str, "address")) {
-			printk("This is a test of memory manager. Warning! You are in kernel mode!\n");
-			printk("Kernel:\n");
-			printk("Start: %x\nEnd: %x\n", (size_t)&kernel_start, (size_t)&kernel_end);
-			getch();
-			printk("stack_top: %x\n", stack_top);
-			printk("&stack_top: %x\n", &stack_top);
-			getch();
-			
-			int *num = (int*) kmalloc(sizeof(int), 0, 0);
-			*num = 0x123ABC;
-			printk("num: %x\n", num);
-			printk("&num: %x\n", &num);
-			printk("*num: %x\n", *num);
-			
-			int snum = 0xABC123;
-			printk("snum: %x\n", snum);
-			printk("&snum: %x\n", &snum);
-			getch();
-			
-			printk("Kernel:\n");
-			printk("Start: %x\nEnd: %x\n", (size_t)&kernel_start, (size_t)&kernel_end);
-			getch();
-			printk("stack_top: %x\n", stack_top);
-			printk("&stack_top: %x\n", &stack_top);
-			getch();
-			
-			kfree(num);
-		} else if (!strcmp(str, "alloc")) {
-			printk("SIZE OF INT: %d\n", sizeof(int));
-			int *a = (int *)kmalloc(sizeof(int), 0, 0);
-			*a = 6;
-			printk("a: %x\n", a);
-			printk("a: %x\n", *a);
-			
-			int *b = (int *)kmalloc(sizeof(int), 0, 0);
-			*b = 5;
-			printk("b: %x\n", b);
-			printk("b: %x\n", *b);
-			kfree(b);
-			
-			int *c = (int *)kmalloc(sizeof(int), 0, 0);
-			*c = 8;
-			printk("c: %x\n", c);
-			printk("c: %x\n", *c);
-			
-			kfree(a);
-			kfree(b);
-			kfree(c);
-			//-------------------------------------------------------
-			char *stmp = (char *)kmalloc(sizeof(char) * 11, 0, 0);
-			memcpy(stmp, "This is.\0", 9);
-			printk(stmp);
-			printk("\n&stmp[0]: %x &stmp[9]: %x\n", stmp, stmp + sizeof(char) * 11);
-			
-			kfree(stmp);
-			
-			char *tmp = (char *)kmalloc(sizeof(char) * 10, 0, 0);
-			memcpy(tmp, "THIS WAS.\0", 10);
-			printk(tmp);
-			printk("\n&tmp[0]: %x &tmp[9]: %x\n", tmp, tmp + sizeof(char) * 11);
-			
-			kfree(tmp);
-		} else if (!strcmp(str, "vga")) {
-			printk("Ab_C\b\b\nNew\tLineWasWritten NUM%x%s%c%c\n", 0x123ABC, "test", 'f', 55);
-			printk("Expected:\nAb\nNew    LineWasWritten NUM0x123abctestf7\n");
-		} else {
-			printk("Command not found!\n");
-		}
-	}
-}
-
 void panic(const char *msg, int line, char *file)
 {
 	cli();
@@ -132,137 +28,4 @@ void reboot()
 		good = inb(0x64);
 	outb(0x64, 0xFE);
 	hlt();
-}
-
-void shell(char *str) {
-	if (!strcmp(str, "exit")) {
-			panic("Exit.", __LINE__, __FILE__);
-	} else if (!strcmp(str, "test")) {
-		testing_shell(str);
-		
-	} else if (!strcmp(str, "time")) {
-		print_time();
-		printk("\n");
-		
-	} else if (!strcmp(str, "tick")) {
-		printk("Tick: %d\n", get_tick());
-	
-	} else if (!strcmp(str, "clock")) {
-		printk("WARN: no clock yet\n");
-	} else if (!strcmp(str, "clear")) {
-		clear_vga();
-		
-	} else if (!strcmp(str, "reset-clock")) {
-		hours = 0;
-		minutes = 0;
-		seconds = 0;
-		
-	} else if (!strcmp(str, "write")) {
-		char *to_write = (char*) kmalloc(sizeof(char) * 30, 0, 0);
-		size_t len = strlen("Let's write this...\n");
-		
-    	memcpy(to_write, "Let's write this...\n", len);
-		call(1, (uint32_t) to_write, (uint32_t) len, 0, 0, 0);
-		kfree(to_write);
-		
-	} else if (!strcmp(str, "reboot")) {
-		reboot();
-		
-	} else if (!strcmp(str, "huge-alloc")) {
-		int *i = (void *) kmalloc(sizeof(int), 0, 0);
-		kfree(i);
-
-	} else if (!strcmp(str, "usermode")) {
-		printk("CPU will jump to usermode, this will stop shell, because it is running only in kernel mode.\n");
-		init_usermode();
-		char s[] = "Welcome in usermode!\n";
-		call(1, (uint32_t) s, (uint32_t) strlen(s), 0, 0, 0);
-	} else if (!strcmp(str, "sys-info")) {
-		printk("%s %s (%s) by %s.\nCopyright C 2015 %s.\nAll rights reserved.\n", OS_Name, Version, Relase_Date, Author, Author);
-		
-	} else if (!strcmp(str, "help")) {
-		printk("After kernel boots up, user must enter user name and machine name. Now, a shell has started. Here user can use some commands:\n");
-		printk("	-> help -> display this page\n");
-		printk("	-> time -> display current time\n");
-		printk("	-> tick -> display current tick\n");
-		printk("	-> clock -> start a clock that can be stoped when ESC is pressed\n");
-		printk("	-> clear -> clear the screen\n");
-		printk("	-> reset-clock -> set clock to 0\n");
-		printk("	-> sys->info -> prints info about system\n");
-		printk("	-> test -> enter into testing shell that allows testing kernel functions\n");
-		printk("	-> exit -> close shell\n\n");
-	
-		printk("Testing shell:\n");
-		printk("	Tests avabile in testing shell:\n");
-		printk("		-> exit -> close testing shell\n");
-		printk("		-> print-stdin -> print all data from standard input\n");
-		printk("		-> size -> print sizes of types, of kernel and RAM\n");
-		printk("		-> page-fault -> generate page fault\n");
-		printk("		-> address -> print addresses of all kernel variables and of some temporary\n");
-		printk("		-> alloc -> test kmalloc and kfree\n");
-		
-	} else {
-		printk("\"%s\" not recognized as internal command or program. Please try again.\n", str);
-	}
-}
-
-void prompt() {
-	char cmd[1024];
-	while (true) {
-		printk("%s@%s:$ ", user, machine);
-		gets(cmd);
-		shell(cmd);
-	}
-}
-
-void welcome() {
-    clear_vga();
-    
-    wstr_color("      |\\     /|(  ____ \\( \\      (  ____ \\(  ___  )(       )(  ____ \\\n", COLOR_RED);
-    wstr_color("      | )   ( || (    \\/| (      | (    \\/| (   ) || () () || (    \\/\n", COLOR_RED);
-    wstr_color("      | | _ | || (__    | |      | |      | |   | || || || || (\\\n", COLOR_RED);
-    wstr_color("      | |( )| ||  __)   | |      | |      | |   | || |(_)| ||  __)\\\n", COLOR_RED);   
-    wstr_color("      | || || || (      | |      | |      | |   | || |   | || (\\\n", COLOR_RED);      
-    wstr_color("      | () () || (____/\\| (____/\\| (____/\\| (___) || )   ( || (____/\\\n", COLOR_RED);
-    wstr_color("      (_______)(_______/(_______/(_______/(_______)|/     \\|(_______/\n\n", COLOR_RED);
-    
-    wstr_color("                           \\__   __/(  ___  )\n", COLOR_YELLOW);
-    wstr_color("                              ) (   | (   ) |\n", COLOR_YELLOW);
-    wstr_color("                              | |   | |   | |\n", COLOR_YELLOW);
-    wstr_color("                              | |   | |   | |\n", COLOR_YELLOW);
-    wstr_color("                              | |   | |   | |\n", COLOR_YELLOW);
-    wstr_color("                              | |   | (___) |\n", COLOR_YELLOW);
-    wstr_color("                              )_(   (_______)\n\n", COLOR_YELLOW);
-    
-    wstr_color(" _______           _______ _________ _______  _        _______  _______ \n", COLOR_BLUE);
-    wstr_color("(  ____ )|\\     /|(  ___  )\\__   __/(  ___  )( (    /|(  ___  )(  ____ \\\\\n", COLOR_BLUE);
-    wstr_color("| (    )|| )   ( || (   ) |   ) (   | (   ) ||  \\  ( || (   ) || (    \\/\n", COLOR_BLUE);
-    wstr_color("| (____)|| (___) || |   | |   | |   | |   | ||   \\ | || |   | || (_____ \n", COLOR_BLUE);
-    wstr_color("|  _____)|  ___  || |   | |   | |   | |   | || (\\ \\) || |   | |(_____  )\n", COLOR_BLUE);
-    wstr_color("| (      | (   ) || |   | |   | |   | |   | || | \\   || |   | |      ) |\n", COLOR_BLUE);
-    wstr_color("| )      | )   ( || (___) |   | |   | (___) || )  \\  || (___) |/\\____) |\n", COLOR_BLUE);
-    wstr_color("|/       |/     \\|(_______)   )_(   (_______)|/    )_)(_______)\\_______)\n", COLOR_BLUE);
-    
-    printk("                                                        by Feraru Mihail");
-    
-    getch();
-    clear_vga();
-}
-
-void login() {
-    wstr_color(" _______           _______ _________ _______  _        _______  _______ \n", COLOR_BLUE);
-    wstr_color("(  ____ )|\\     /|(  ___  )\\__   __/(  ___  )( (    /|(  ___  )(  ____ \\\\\n", COLOR_BLUE);
-    wstr_color("| (    )|| )   ( || (   ) |   ) (   | (   ) ||  \\  ( || (   ) || (    \\/\n", COLOR_BLUE);
-    wstr_color("| (____)|| (___) || |   | |   | |   | |   | ||   \\ | || |   | || (_____ \n", COLOR_BLUE);
-    wstr_color("|  _____)|  ___  || |   | |   | |   | |   | || (\\ \\) || |   | |(_____  )\n", COLOR_BLUE);
-    wstr_color("| (      | (   ) || |   | |   | |   | |   | || | \\   || |   | |      ) |\n", COLOR_BLUE);
-    wstr_color("| )      | )   ( || (___) |   | |   | (___) || )  \\  || (___) |/\\____) |\n", COLOR_BLUE);
-    wstr_color("|/       |/     \\|(_______)   )_(   (_______)|/    )_)(_______)\\_______)\n", COLOR_BLUE);
-    
-    printk("Log in please.\n");
-    printk("Username: ");
-    gets(user);
-    printk("Machine: ");
-    gets(machine);
-    printk("Loged in!\n");
 }
